Changes Strategy for Splitting Up Script
This commit modifies the way the script is split up, opting for a `while loop` in order to use an `index` variable for keeping track of lines. This allows for "looking ahead" at the next line as well as incrementing the index variable at different stages of the search step.

From a quick glance, it seems that it doesn't work perfectly but it's on it's way.